Cleanup types and stories
