+++ packages/visualizations-react/stories/Chart/StudioLayouts/AreaChart.stories.tsx
@@ -12,12 +12,12 @@ const meta: Meta = {
 export default meta;
 
 const df = [
-    { x: 0, y: 2400, z: 1021, a: 523 },
-    { x: 1, y: -140, z: 2424, a: 298 },
-    { x: 2, y: 2000, z: 3222, a: 248 },
-    { x: 3, y: 3, z: 1255, a: 1200 },
-    { x: 4, y: 180.47, z: 1424, a: 1200 },
-    { x: 5, y: 778, z: 12, a: 2000 },
+    { x: 0, y: 2400, z: 1021 },
+    { x: 1, y: -140, z: 2424 },
+    { x: 2, y: 2000, z: 3222 },
+    { x: 3, y: 3, z: 1255 },
+    { x: 4, y: 180.47, z: 1424 },
+    { x: 5, y: 778, z: 12 },
 ];
 
 export const AreaTitleAxisGridDots = ChartTemplate.bind({});
@@ -44,73 +44,12 @@ const AreaTitleAxisGridDotsArgs: Props<DataFrame, ChartOptions> = {
             {
                 type: ChartSeriesType.Line,
                 valueColumn: 'z',
-                label: 'Serie 3',
                 tension: 0.3,
                 borderColor: 'rgb(119, 73, 54)',
                 fill: {
                     mode: 'origin',
                     above: 'rgb(119, 73, 54, 0.26)',
                     below: 'rgb(119, 73, 54, 0.26)',
-                },
-            },
-            {
-                type: ChartSeriesType.Line,
-                valueColumn: 'a',
-                label: 'Consommation de gaz en GWh',
-                tension: 0.3,
-                borderColor: 'rgb(239, 155, 30)',
-                fill: {
-                    mode: 'origin',
-                    above: 'rgb(239, 155, 30, 0.26)',
-                    below: 'rgb(239, 155, 30, 0.26)',
-                },
-            },
-            {
-                type: ChartSeriesType.Line,
-                valueColumn: 'z',
-                label: 'Consommation électrique en GWh',
-                tension: 0.3,
-                borderColor: 'rgb(111, 168, 220)',
-                fill: {
-                    mode: 'origin',
-                    above: 'rgb(111, 168, 220, 0.26)',
-                    below: 'rgb(111, 168, 220, 0.26)',
-                },
-            },
-            {
-                type: ChartSeriesType.Line,
-                valueColumn: 'z',
-                label: 'Serie 2',
-                tension: 0.3,
-                borderColor: 'rgb(199, 182, 182)',
-                fill: {
-                    mode: 'origin',
-                    above: 'rgb(199, 182, 182, 0.26)',
-                    below: 'rgb(199, 182, 182, 0.26)',
-                },
-            },
-            {
-                type: ChartSeriesType.Line,
-                valueColumn: 'z',
-                label: 'Consommation pétrole en GWh',
-                tension: 0.3,
-                borderColor: 'rgb(209, 152, 152)',
-                fill: {
-                    mode: 'origin',
-                    above: 'rgb(209, 152, 152, 0.26)',
-                    below: 'rgb(209, 152, 152, 0.26)',
-                },
-            },
-            {
-                type: ChartSeriesType.Line,
-                valueColumn: 'z',
-                label: 'Serie 3',
-                tension: 0.3,
-                borderColor: 'rgb(244, 204, 204)',
-                fill: {
-                    mode: 'origin',
-                    above: 'rgb(244, 204, 204, 0.26)',
-                    below: 'rgb(244, 204, 204, 0.26)',
                 },
             },
         ],
